requirements.txt
django
jupyter
requests
Celery
redis
django-celery-beat
django-celery-results
python-decouple
pytz
openai